+++ libraries/typescript/packages/client/src/index-browser.ts
@@ -37,10 +37,8 @@ export {
 export {
   telFetch,
   capturePostHog,
-  captureScarf,
   POSTHOG_HOST,
   POSTHOG_API_KEY,
-  SCARF_GATEWAY_URL,
 } from "./telemetry/tel-fetch.js";
 export { getPackageVersion, VERSION } from "./utils/version.js";
 export { detectFavicon } from "./utils/favicon.js";